feat(template): rename Infra to Iac in the TodoTemplate project #1948 (#2045)
