+++ src/Tooling/Bit.Tooling.Templates/TodoTemplate/Iac/Pulumi.yaml
@@ -1,0 +1,3 @@
+﻿name: TodoTemplate.Iac
+runtime: dotnet
+description: TodoTemplate infrastructure as a code
