--- src/Tooling/Bit.Tooling.Templates/TodoTemplate/Infra/Pulumi.yaml
@@ -1,3 +1,0 @@
-﻿name: TodoTemplate.Infra
-runtime: dotnet
-description: TodoTemplate infra as a code
